--- src/main/java/use_case/previousPlan/PreviousPlanInteractor.java
@@ -1,4 +1,0 @@
-package use_case.previousPlan;
-
-public class PreviousPlanInteractor {
-}
