+++ src/main/java/use_case/previous_plan/PreviousPlanInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.previousPlan;
+package use_case.previous_plan;
 
 public interface PreviousPlanInputBoundary {
     void execute(PreviousPlanInputData previousPlanInputData);
