+++ src/main/java/use_case/previous_plan/PreviousPlanInputData.java
@@ -1,4 +1,4 @@
-package use_case.previousPlan;
+package use_case.previous_plan;
 
 import entity.User;
